Lesson_29/weather_bot/requirements.txt
python-dotenv==1.2.1
python-telegram-bot==20.3
requests==2.32.5
